requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.47
Flask-JWT-Extended==4.4.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
Flask-CORS==3.1.1